2020/10/20

+++ src/main/java/com/chat/talk/repository/FilesRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.chat.talk.model.Files;
 
-public interface  FilesRepository extends JpaRepository<Files, Integer> {
+public interface  FilesRepository extends JpaRepository<Files, Long> {
 
-	Files findByFno(int fno);
+	Files findByFno(Long fno);
 }
